// apps/backend/src/common/enums/error-code.enum.ts
export enum ErrorCode {
  // Authentication Errors (AUTH_XXX)
  AUTH_UNAUTHORIZED = 'AUTH_001',
  AUTH_INVALID_TOKEN = 'AUTH_002',
  AUTH_TOKEN_EXPIRED = 'AUTH_003',
  AUTH_MISSING_CREDENTIALS = 'AUTH_004',
  AUTH_INVALID_CREDENTIALS = 'AUTH_005',
  AUTH_EMAIL_NOT_VERIFIED = 'AUTH_006',
  AUTH_USER_DISABLED = 'AUTH_007',

  // System/General Errors (SYS_XXX)
  SYS_INTERNAL_ERROR = 'SYS_001',
  SYS_SERVICE_UNAVAILABLE = 'SYS_002',
  SYS_EXTERNAL_SERVICE_TIMEOUT = 'SYS_003',
  SYS_VALIDATION_FAILED = 'SYS_004',
  SYS_NOT_FOUND = 'SYS_005',
  SYS_FORBIDDEN = 'SYS_006',
  SYS_BAD_REQUEST = 'SYS_007',
  SYS_RATE_LIMIT_EXCEEDED = 'SYS_008',
  SYS_CONFLICT = 'SYS_009',
  SYS_UNSUPPORTED_MEDIA_TYPE = 'SYS_010',
  SYS_PAYLOAD_TOO_LARGE = 'SYS_011',

  // User Errors (USER_XXX)
  USER_NOT_FOUND = 'USER_001',
  USER_ALREADY_EXISTS = 'USER_002',
  USER_INVALID_DATA = 'USER_003',

  // Stellar/Blockchain Errors (STEL_XXX)
  STEL_WALLET_NOT_FOUND = 'STEL_001',
  STEL_INSUFFICIENT_FUNDS = 'STEL_002',
  STEL_TRANSACTION_FAILED = 'STEL_003',
  STEL_INVALID_ADDRESS = 'STEL_004',
  STEL_SIMULATION_FAILED = 'STEL_005',
  STEL_RPC_TIMEOUT = 'STEL_006',
  STEL_RPC_UNAVAILABLE = 'STEL_007',
  STEL_FRIENDBOT_ALREADY_FUNDED = 'STEL_008',
  STEL_FRIENDBOT_FAILED = 'STEL_009',
  STEL_TESTNET_ONLY = 'STEL_010',

  // Treasury Stream Errors (TREAS_XXX)
  TREAS_NOT_CONFIGURED = 'TREAS_001',
  TREAS_NOT_INITIALIZED = 'TREAS_002',
  TREAS_UNAUTHORIZED = 'TREAS_003',
  TREAS_INVALID_AMOUNT = 'TREAS_004',
  TREAS_INVALID_DURATION = 'TREAS_005',
  TREAS_INVALID_START_TIME = 'TREAS_006',
  TREAS_STREAM_NOT_FOUND = 'TREAS_007',
  TREAS_NOTHING_TO_CLAIM = 'TREAS_008',
  TREAS_REENTRANCY = 'TREAS_009',
  TREAS_RPC_UNAVAILABLE = 'TREAS_010',
  TREAS_TRANSACTION_FAILED = 'TREAS_011',

  // Vesting Wallet Errors (VEST_XXX)
  VEST_NOT_CONFIGURED = 'VEST_001',
  VEST_NOT_INITIALIZED = 'VEST_002',
  VEST_UNAUTHORIZED = 'VEST_003',
  VEST_INVALID_AMOUNT = 'VEST_004',
  VEST_INVALID_DURATION = 'VEST_005',
  VEST_INVALID_START_TIME = 'VEST_006',
  VEST_NOT_FOUND = 'VEST_007',
  VEST_NOTHING_TO_CLAIM = 'VEST_008',
  VEST_REENTRANCY = 'VEST_009',
  VEST_RPC_UNAVAILABLE = 'VEST_010',
  VEST_TRANSACTION_FAILED = 'VEST_011',

  // News & Analytics Errors (ANLY_XXX)
  ANLY_SENTIMENT_ANALYSIS_FAILED = 'ANLY_001',
  ANLY_NEWS_SYNC_FAILED = 'ANLY_002',
  ANLY_DATA_NOT_FOUND = 'ANLY_003',
}